Refactor: split websockets.py into websockets.py and websockets_base.py
